@corona10 2020-03-18
The benchmark from the https://bugs.python.org/issue37207#msg364428 was executed by below script.

```
import pyperf

runner = pyperf.Runner()
runner.timeit(name="frozenset create test",
              stmt="frozenset([1, 2, 3, 4, 5])")

```