@corona10 2020-03-18
@vstinner 

Updated, During I work on helper function, I updated my git client and 
during this process, it makes duplicated commits 27db91d d467294 .
I resolved all confliction and now it works okay. I also checked reference leak
through `./python.exe -m test test_set -R 3:3`and there was no leak ;-)