@corona10 2020-03-18
And performance improvement is still valid.

```
Mean +- std dev: [master] 2.30 us +- 0.10 us -> [pr] 2.10 us +- 0.06 us: 1.10x faster (-9%)
```