@vstinner 2020-03-18
Thanks @corona10: 1.09x faster (-9%) is nice to have (-200 ns). And IMHO the change makes the code more readable with new comments.